+++ src/sisl/physics/_matrix_dk.pyx
@@ -21,16 +21,17 @@ __all__ = ["matrix_dk", "matrik_dk_nc", "matrik_dk_nc_diag", "matrik_dk_so"]
 
 def _phase_dk(gauge, M, sc, np.ndarray[np.float64_t, ndim=1, mode='c'] k, dtype):
     # dtype *must* be passed through phase_dtype
+    gauge = {"R": "cell", "r": "orbital", "orbitals": "orbital"}.get(gauge, gauge)
 
     # This is the differentiated matrix with respect to k
     # See _phase.pyx, we are using exp(i k.R/r)
     #  i R
-    if gauge == 'R':
+    if gauge == 'cell':
         iRs = phase_rsc(sc, k, dtype).reshape(-1, 1)
         iRs = (1j * _dot(sc.sc_off, sc.cell) * iRs).astype(dtype, copy=False)
         p_opt = 1
 
-    elif gauge == 'r':
+    elif gauge == 'orbital':
         M.finalize()
         rij = M.Rij()._csr._D
         iRs = (1j * rij * phase_rij(rij, sc, k, dtype).reshape(-1, 1)).astype(dtype, copy=False)