+++ src/sisl/viz/processors/eigenstate.py
@@ -178,8 +178,8 @@ def project_wavefunction(
     """
     grid = create_wf_grid(eigenstate, grid_prec=grid_prec, grid=grid, geometry=geometry)
 
-    # Ensure we are dealing with the R gauge
-    eigenstate.change_gauge("R")
+    # Ensure we are dealing with the cell gauge
+    eigenstate.change_gauge("cell")
 
     # Finally, insert the wavefunction values into the grid.
     sisl.physics.electron.wavefunction(